requirements.txt

pandas==2.1.3
numpy==1.24.3

torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
scikit-learn==1.3.0

faiss-cpu==1.7.4

requests==2.31.0
aiohttp==3.8.4
python-telegram-bot==20.4
vk-api==11.9.0

pydantic==1.10.12
python-dotenv==1.0.0
loguru==0.7.0
tqdm==4.65.0
Jinja2==3.1.2